backend/requirements.txt
faiss-cpu
transformers
sentence_transformers
sqlalchemy
fastapi
uvicorn
rank_bm25
atlassian-python-api
markdown
beautifulsoup4
html2text
python-dotenv